pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memagent"
version = "0.3.8"
description = "Open-source cross-agent memory layer for AI coding agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Open Source Community"},
]
keywords = ["ai", "agents", "memory", "claude", "codex", "cursor", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "pyyaml>=6.0",
]
embeddings = [
    "sentence-transformers>=3.0",
]
mcp = [
    "fastmcp>=3.0",
]
dashboard = [
    "fastapi>=0.110",
    "uvicorn>=0.25",
]
postgres = [
    "psycopg>=3.0",
]
all = [
    "sentence-transformers>=3.0",
    "fastmcp>=3.0",
    "fastapi>=0.110",
    "uvicorn>=0.25",
    "psycopg>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
memagent = "memagent.cli:main"
mg = "memagent.cli:main"

[project.urls]
Homepage = "https://github.com/Vish150988/agentmemory"
Repository = "https://github.com/Vish150988/agentmemory"
Issues = "https://github.com/Vish150988/agentmemory/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"memagent/dashboard.py" = ["E501"]
"memagent/shell.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
